refactor(cli): extract SessionWatcher deep module from LiveScanStore
LiveScanStore had grown to 1429 lines, mixing scan orchestration with the
entire file-watch subsystem: recursive-watch detection, non-recursive
fallback trees, write-stability polling, and per-agent debouncing. Extract
that subsystem into a SessionWatcher deep module with a tiny interface:

  start(agentNames) / dispose() / onAgentsChanged(cb)

The watcher owns watchers/fallbackWatchScopes/stablePaths and the
write-stability polling internally. The store registers an
onAgentsChanged listener that runs the refresh-scheduling logic
(pendingRefreshPathCounts bump + empty-agent debounce + scheduleRefresh),
which stays in the store because it depends on byAgent session counts.

resolveAgentWatchTargets and the watch helpers move with the watcher;
live-scan re-exports resolveAgentWatchTargets so existing test imports are
unchanged. The fs.watch mock lives at the node:fs layer, so the watch
test cases in live-scan-store.test.ts drive the watcher through the same
fsWatch.watchers array as before.
